LayoutTests/imported/w3c/web-platform-tests/scroll-animations/animation-trigger/trigger-scope-named.tentative-expected.txt

In-scope Target
SOURCE
Out-of-scope Target

FAIL target within named trigger-scope finds trigger within scope; out-of-scope target does not. assert_equals: animation on in_scope_target has playState paused. expected "paused" but got "running"
